pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.23.0",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "psutil>=7.0.0",
    "scikit-learn>=1.7.0",
    "streamlit>=1.46.1",
]
